veil

+++ 9.Maintain-Access/1.OS-Backdoors/0.MultiOS-Backdoor/veil/#veil-update.sh
@@ -1,19 +1,22 @@
 #!/bin/bash
 
-
-# Veil Evasion is no longer supported, use Veil 3.0! https://github.com/Veil-Framework/Veil
-
-
 bold=$(tput bold)
 normal=$(tput sgr0)
 
 GITREPO=https://github.com/Veil-Framework/Veil
 BRANCH=master
 GITREPOROOT=/opt/ITSEC/9.Maintain-Access/1.OS-Backdoors/0.MultiOS-Backdoor/Veil/Veil-Framework/Veil
+GITCONFDIR=/opt/ITSEC/9.Maintain-Access/1.OS-Backdoors/0.MultiOS-Backdoor/Veil/Veil-Framework/Veil/.git
 GITCLONEDIR=/opt/ITSEC/9.Maintain-Access/1.OS-Backdoors/0.MultiOS-Backdoor/Veil/Veil-Framework
 EXECUTEABLE1=Veil.py
 EXECUTEABLE2=Veil
 BINDIR=/usr/local/bin
+GITRESET () {
+	git clean -f
+	git fetch origin
+	git reset --hard origin/$BRANCH
+	git pull
+}
 APTLSTDIR=/opt/ITSEC-Install-Scripts/0.Initial/lst/apt
 GITSBMDLINIT () {
 	git submodule init
@@ -27,26 +30,54 @@ __     _______ ___ _
  \ \ / /|  _|  | || |    
   \ V / | |___ | || |___ 
    \_/  |_____|___|_____|
-                         
-INSTALL
+                            
+UPDATE
 ${normal}"
+
+if [ ! -d $GITCONFDIR ]
+
+then
 
 mkdir -p $GITCLONEDIR
 cd $GITCLONEDIR
 git clone -b $BRANCH $GITREPO
+
+else
+
+echo "${bold}REPO EXISTS, skip clone...${normal}"
+
+fi
+
 cd $GITREPOROOT
 
-### DEPS:
-# no deps noted, feel free to add ...
-### DEPS END
+if git checkout $BRANCH &&
+    git fetch origin $BRANCH &&
+    [ `git rev-list HEAD...origin/$BRANCH --count` != 0 ] &&
+    git merge origin/$BRANCH
+then
+    
+cd $GITREPOROOT
 
+
+cd 
+GITRESET
 GITSBMDLINIT
 
-sed -i 's/wine-stable/wine/' /opt/ITSEC/9.Maintain-Access/1.OS-Backdoors/0.MultiOS-Backdoor/Veil/Veil-Framework/Veil/setup/setup.sh
+sed -i 's/wine-stable/wine/' /opt/ITSEC/9.Maintain-Access/2.Generate/Veil/Veil-Framework/Veil/setup/setup.sh
 cd setup
-./setup.sh -s
-
-chmod +x $EXECUTEABLE1
+./setup.sh -c
 
 sudo rm -f $BINDIR/$EXECUTEABLE2
 sudo ln -s $GITREPOROOT/$EXECUTEABLE1 $BINDIR/$EXECUTEABLE2
+
+echo "${bold}
+UPDATED
+${normal}"
+
+else
+
+echo "${bold}
+UP TO DATE
+${normal}"
+	
+fi
